Move tests that check actual preseeding from cmd/snap-preseed to image/preseed. Update remaining tests with proper mocks.
